+++ server/test/unit/Ogmios/Data/JsonSpec.hs
@@ -32,11 +32,15 @@ import Ogmios.Data.Json
     , decodeWith
     , encodeAcquireFailure
     , encodeBlock
+    , encodeExUnits
     , encodePoint
+    , encodeScriptFailure
     , encodeSubmitTxError
     , encodeTip
     , encodeTxId
+    , encodeTxIn
     , jsonToByteString
+    , stringifyRdmrPtr
     )
 import Ogmios.Data.Json.Orphans
     ()
@@ -122,7 +126,11 @@ import Ogmios.Data.Protocol.TxMonitor
     , _encodeSizeAndCapacityResponse
     )
 import Ogmios.Data.Protocol.TxSubmission
-    ( SubmitTxResponse, _encodeSubmitTxResponse )
+    ( EvaluateTxResponse
+    , SubmitTxResponse (..)
+    , _encodeEvaluateTxResponse
+    , _encodeSubmitTxResponse
+    )
 import Ouroboros.Consensus.Cardano.Block
     ( CardanoEras, GenTx, HardForkApplyTxErr (..) )
 import Ouroboros.Network.Block
@@ -141,6 +149,7 @@ import Test.Generators
     , genCompactGenesisResult
     , genDelegationAndRewardsResult
     , genEpochResult
+    , genEvaluateTxResponse
     , genHardForkApplyTxErr
     , genInterpreterResult
     , genMempoolSizeAndCapacity
@@ -155,6 +164,7 @@ import Test.Generators
     , genProposedPParamsResult
     , genRewardInfoPoolsResult
     , genRewardProvenanceResult
+    , genSubmitResult
     , genSystemStart
     , genTip
     , genTxId
@@ -274,14 +284,20 @@ spec = do
             "RequestNextResponse_1.json"
             "ogmios.wsp.json#/properties/RequestNextResponse"
 
-    context "validate tx submission request/response against JSON-schema" $ do
+    context "validate tx-submission request/response against JSON-schema" $ do
         prop "deserialise signed transactions" prop_parseSubmitTx
 
         validateToJSON
             (arbitrary @(Wsp.Response (SubmitTxResponse Block)))
-            (_encodeSubmitTxResponse (Proxy @Block) encodeSubmitTxError)
+            (_encodeSubmitTxResponse (Proxy @Block) encodeTxId encodeSubmitTxError)
             (200, "TxSubmission/Response/SubmitTx")
             "ogmios.wsp.json#/properties/SubmitTxResponse"
+
+        validateToJSON
+            (arbitrary @(Wsp.Response (EvaluateTxResponse Block)))
+            (_encodeEvaluateTxResponse (Proxy @Block) stringifyRdmrPtr encodeExUnits encodeScriptFailure encodeTxIn)
+            (100, "TxSubmission/Response/EvaluateTx")
+            "ogmios.wsp.json#/properties/EvaluateTxResponse"
 
         goldenToJSON
             "SubmitTxResponse_1.json"
@@ -572,8 +588,18 @@ instance Arbitrary RequestNext where
     shrink = genericShrink
     arbitrary = reasonablySized genericArbitrary
 
-instance Arbitrary (SubmitResult (HardForkApplyTxErr (CardanoEras StandardCrypto))) where
+instance Arbitrary (SubmitTxResponse Block) where
+    shrink = genericShrink
+    arbitrary = genericArbitrary
+
+instance Arbitrary (HardForkApplyTxErr (CardanoEras StandardCrypto)) where
     arbitrary = genHardForkApplyTxErr
+
+instance Arbitrary (SubmitResult (HardForkApplyTxErr (CardanoEras StandardCrypto))) where
+    arbitrary = genSubmitResult
+
+instance Arbitrary (EvaluateTxResponse Block) where
+    arbitrary = genEvaluateTxResponse
 
 instance Arbitrary (Acquire Block) where
     shrink = genericShrink